correcting a bug before changing how the VM works (an example code using segments wasn't running because we were trying to getXBytes(bytecode, byteCount=m_ip)

+++ value.hpp
@@ -39,7 +39,7 @@ namespace kafe
     {
         ValueType type;
 
-        int intValue;
+        long intValue;
         double doubleValue;
         bool boolValue;
         std::string stringValue;
